backend/requirements.txt
flask
tensorflow-cpu
pandas
numpy
scikit-learn
nltk
aif360
gunicorn
flask-cors
